refactor: Move generic button to its own file
