+++ app/src/main/java/com/infomaniak/swisstransfer/ui/components/Buttons.kt
@@ -20,15 +20,11 @@ package com.infomaniak.swisstransfer.ui.components
 import android.content.res.Configuration
 import androidx.annotation.StringRes
 import androidx.compose.foundation.layout.*
-import androidx.compose.material3.*
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
@@ -37,7 +33,6 @@ import androidx.compose.ui.unit.dp
 import com.infomaniak.swisstransfer.R
 import com.infomaniak.swisstransfer.ui.images.AppImages.AppIcons
 import com.infomaniak.swisstransfer.ui.images.icons.Add
-import com.infomaniak.swisstransfer.ui.theme.CustomShapes
 import com.infomaniak.swisstransfer.ui.theme.Dimens
 import com.infomaniak.swisstransfer.ui.theme.Margin
 import com.infomaniak.swisstransfer.ui.theme.SwissTransferTheme
@@ -121,98 +116,12 @@ private fun CoreTextButton(
 }
 
 @Composable
-fun GenericButton(
-    modifier: Modifier,
-    style: ButtonType = ButtonType.PRIMARY,
-    enabled: () -> Boolean = { true },
-    showIndeterminateProgress: () -> Boolean = { false },
-    progress: (() -> Float)? = null,
-    onClick: () -> Unit,
-    contentPadding: PaddingValues = ButtonDefaults.ContentPadding,
-    content: @Composable () -> Unit,
-) {
-    val isEnabled by remember(progress) { derivedStateOf { enabled() && !showIndeterminateProgress() && progress == null } }
-    val buttonColors = style.buttonColors()
-
-    Button(
-        modifier = modifier,
-        colors = buttonColors,
-        shape = CustomShapes.MEDIUM,
-        enabled = isEnabled,
-        contentPadding = contentPadding,
-        onClick = onClick,
-    ) {
-        when {
-            progress != null -> {
-                val (progressColor, progressModifier) = getProgressSpecs(buttonColors)
-                KeepButtonSize(content) {
-                    CircularProgressIndicator(modifier = progressModifier, color = progressColor, progress = progress)
-                }
-            }
-            showIndeterminateProgress() -> {
-                val (progressColor, progressModifier) = getProgressSpecs(buttonColors)
-                KeepButtonSize(content) {
-                    CircularProgressIndicator(modifier = progressModifier, color = progressColor)
-                }
-            }
-            else -> content()
-        }
-    }
-}
-
-@Composable
-fun KeepButtonSize(targetSizeContent: @Composable () -> Unit, content: @Composable () -> Unit) {
-    Box(contentAlignment = Alignment.Center) {
-        Row(modifier = Modifier.alpha(0f)) {
-            targetSizeContent()
-        }
-        content()
-    }
-}
-
-@Composable
 private fun ButtonTextContent(imageVector: ImageVector?, titleRes: Int) {
     imageVector?.let {
         Icon(modifier = Modifier.size(Dimens.SmallIconSize), imageVector = it, contentDescription = null)
         Spacer(Modifier.width(Margin.Mini))
     }
     Text(text = stringResource(id = titleRes), style = SwissTransferTheme.typography.bodyMedium)
-}
-
-@Composable
-private fun getProgressSpecs(buttonColors: ButtonColors): Pair<Color, Modifier> {
-    val progressColor = buttonColors.disabledContentColor
-    val progressModifier = Modifier
-        .fillMaxHeight(0.8f)
-        .aspectRatio(1f)
-    return Pair(progressColor, progressModifier)
-}
-
-enum class ButtonType(val buttonColors: @Composable () -> ButtonColors) {
-    PRIMARY({
-        ButtonDefaults.buttonColors(
-            containerColor = SwissTransferTheme.materialColors.primary,
-            contentColor = SwissTransferTheme.materialColors.onPrimary,
-        )
-    }),
-    SECONDARY({
-        ButtonDefaults.buttonColors(
-            containerColor = SwissTransferTheme.colors.tertiaryButtonBackground,
-            contentColor = SwissTransferTheme.materialColors.primary,
-        )
-    }),
-    TERTIARY({
-        ButtonDefaults.buttonColors(
-            containerColor = Color.Transparent,
-            contentColor = SwissTransferTheme.materialColors.primary,
-        )
-    }),
-    ERROR({
-        ButtonDefaults.buttonColors(
-            containerColor = SwissTransferTheme.materialColors.error,
-            contentColor = SwissTransferTheme.materialColors.onError,
-        )
-    }),
 }
 
 private enum class ButtonSize(val height: Dp) {
